requirements.txt
flask==2.3.3
boto3==1.28.44
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3  
pandas==2.1.0
